hf5

+++ routes/news.js
@@ -1,37 +1,31 @@
-/**
- * Require all the middlewares
- */
+
 var authMW = require('../middlewares/general/auth');
 var renderMW = require('../middlewares/general/render');
-var mainRedirectMW = require('../middlewares/general/mainredirect');
+var requirePermissionMW = require('../middlewares/general/requirePermission');
 
 var updateNewsMW = require('../middlewares/news/updatenews');
 var deleteNewsMW = require('../middlewares/news/deletenews');
-var getNewsList = require('../middlewares/news/getnewslist');
+var getNewsListMW = require('../middlewares/news/getnewslist');
+var getNewsMW = require('../middlewares/news/getnews');
+
+var NewsModel = require('../models/news');
+var UserModel = require('../models/user');
 
 module.exports = function (app) {
 
     var objectRepository = {
+        newsModel: NewsModel,
+        userModel: UserModel
     };
-
-    /**
-     * The page for new news
-     *      GET:    The page with the form where the admin give the datas
-     *      POST:   Save the new new
-     *                  -> than redirects to the news' page (/news)
-     */
-    app.use('/news/add',
-        authMW(objectRepository),
-        updateNewsMW(objectRepository),
-        renderMW(objectRepository, 'addnews')
-    );
 
     /**
      * Deletes the news with id ':id'
      *      -> than redirects to news' page (/news)
      */
-    app.use('/news/del/:id',
+    app.use('/news/:id/del',
         authMW(objectRepository),
+        requirePermissionMW(objectRepository, 'admin'),
+        getNewsMW(objectRepository),
         deleteNewsMW(objectRepository)
     );
 
@@ -39,20 +33,12 @@ module.exports = function (app) {
      * Updates the news with id ':id'
      *      -> than redirects to news' page (/news)
      */
-    app.use('/news/mod/:id',
-        function(req,res,next){
-            var news={
-                    id: 1,
-                    title: '1 - A hír címe',
-                    duedate: '2016.05.06',
-                    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
-                };
-            res.tpl.news = news;
-            next();
-        },
+    app.use('/news/:id/mod',
         authMW(objectRepository),
+        requirePermissionMW(objectRepository, 'admin'),
+        getNewsMW(objectRepository),
         updateNewsMW(objectRepository),
-        renderMW(objectRepository,'addnews')
+        renderMW(objectRepository,'newsUpdate')
     );
 
     /**
@@ -61,40 +47,9 @@ module.exports = function (app) {
      *      admins can delete and modify the news here
      */
     app.use('/news',
-        function(req,res,next){
-          var news= [
-              {
-                  id: 1,
-                  title: '1 - A hír címe',
-                  duedate: '2016.05.06',
-                  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
-              },
-              {
-                  id: 2,
-                  title: '2 - A hír címe',
-                  duedate: '2016.05.06',
-                  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
-              },
-              {
-                  id: 3,
-                  title: '3 - A hír címe',
-                  duedate: '2016.05.06',
-                  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.'
-              }
-          ];
-          res.tpl.news = news;
-          next();
-        },
         authMW(objectRepository),
-        getNewsList(objectRepository),
-        renderMW(objectRepository,'news')
-    );
-
-    /**
-     *
-     */
-    app.use('/',
-        mainRedirectMW(objectRepository)
+        getNewsListMW(objectRepository),
+        renderMW(objectRepository,'newsAll')
     );
 
 };
